Fix some datatype errors from previous commits

+++ tests/alltypes_proto3_callback/decode_alltypes_callback.c
@@ -11,7 +11,7 @@
 #include "test_helpers.h"
 
 #define TEST(x) if (!(x)) { \
-    printf("Test %s failed (in field %d).\n", #x, field->tag); \
+    printf("Test %s failed (in field %d).\n", #x, (int)field->tag); \
     return false; \
     }
